Initial import of acs-0.25
ACS is a general purpose circuit simulator.  It performs nonlinear
dc and transient analyses, fourier analysis, and ac analysis
linearized at an operating point.  It is fully interactive and
command driven.  It can also be run in batch mode or as a server.
The output is produced as it simulates.  Spice compatible models
for the MOSFET (level 1,2,3,6) and diode are included in this
release.

Since it is fully interactive, it is possible to make changes and
re-simulate quickly.  The interactive design makes it well suited
to the typical iterative design process used it optimizing a circuit
design.  It is also well suited to undergraduate teaching where
Spice in batch mode can be quite intimidating.  This version, while
still officially in beta test, should be stable enough for basic
undergraduate teaching and courses in MOS design, but not for
bipolar design.

In batch mode it is mostly Spice compatible, so it is often possible
to use the same file for both ACS and Spice.