Initial import of gopher-3.0.2, a distributed document delivery client and server, into the NetBSD Packages Collection.
The Internet Gopher is a distributed document delivery service.  It
allows a neophyte user to access various types of data residing on
multiple hosts in a seamless fashion.  This is accomplished by
presenting the user a hierarchical arrangement of documents and by
using a client-server communications model.  The Internet Gopher
Server accepts simple queries, and responds by sending the client a
document.

This is the Gopher+ distribution based on the University of
Minnesota's Internet Gopher+ distribution for UNIX machines.
University of Minnesota actively developed this software until 1995.
In 2000, they released it to the public under the GNU General Public
License.  This represents the latest version of their tree, as adopted
by a group of programmers.  A client for VMS is included but is not
currently supported.

Provided in PR 15285 by jgoerzen@complete.org