Initial import of ROX-wallpaper 0.1.1 to the NetBSD Packages Collection.
This ROX application can be used to place an image on your desktop
background. You can drag an image onto the Wallpaper icon to set that
image as the desktop background directly, or you can click on the
Wallpaper icon to set things up for automatic use.

Provided by Rui-Xiang Guo in PR#17042. Thanks!