Initial import of xmlcatmgr, version 0.1:
xmlcatmgr is a command line script used to manipulate SGML and XML catalogs.
It is designed with simplicity in mind:  it does not depend on external
programs nor libraries, so it is ideal to manage catalog files from inside
pkgsrc.

XML catalogs are files that contain mappings from public identifiers to
system identifiers.  This allows XML parsers to locate public identifiers
in the local system.