Import of gltron-0.62: 3D lightcycle game inspired by the movie Tron
You steer a futuristic bike, called lightcycle.  Combat takes place in
a rectangular arena.  Your bike leaves a trail behind, which is like
a wall.  The goal is to force the other players to drive into a wall.
The winner is the last player alive.

As the name suggests, the game is inspired by the movie Tron. Or more
exactly, by all the games that where inspired by the movie Tron.

Package submitted by Quentin Garnier via pkgsrc-wip.