Initial import of freeglut-2.2.0, from pkgsrc-wip.
freeglut is a completely OpenSourced alternative to the OpenGL
Utility Toolkit (GLUT) library. GLUT was originally written by
Mark Kilgard to support the sample programs in the second edition
OpenGL 'RedBook'. Since then, GLUT has been used in a wide variety
of practical applications because it is simple, universally available
and highly portable.