Initial import of websvn-1.39 from pkgsrc-wip.
WebSVN offers a view onto your subversion repositories that's been designed to
reflect the Subversion methodology. You can view the log of any file or
directory and see a list of all the files changed, added or deleted in any
given revision. You can also view the differences between 2 versions of a
file so as to see exactly what was changed in a particular revision.

WebSVN offers the following features:

    * Easy to use interface
    * Highly customisable templating system
    * Log message searching
    * Colourisation of file listings
    * Fast browsing thanks to internal caching feature
    * Apache MultiViews support

Since it's written using PHP, WebSVN is also very portable and easy to
install.