Initial import of memtester-2.93.1 from pkgsrc-wip, contributed by <pancake at phreaker.net> with some minor changes by me.
memtest is a utility for testing the memory subsystem in a computer
to determine if it is faulty. The original source was by Simon
Kirby <sim@stormix.com>. I have by this time completely rewritten
the original source, and added many additional tests to help catch
borderline memory. I also rewrote the original tests (which catch
mainly memory bits which are stuck permanently high or low) so that
they run approximately an order of magnitude faster.