Initial import of Pyro from pkgsrc-wip.
Pyro is an acronym for PYthon Remote Objects. It is an advanced and powerful
Distributed Object Technology system written entirely in Python, that is
designed to be very easy to use. It resembles Java's Remote Method
Invocation (RMI). It is less similar to CORBA - which is a system- and
language independent Distributed Object Technology and has much more to
offer than Pyro or RMI. But Pyro is small, simple and free (MIT software
license).

Get the manual from

	http://pyro.sourceforge.net/pyro-manual/PyroManual.html

Check quite impressive success stories on

	http://pyro.sourceforge.net/projects.html