Initial import of dbmail-2.0.1.
Dbmail is the name of a group of programs that enable the possiblilty
of storing and retrieving mail messages from a database. Currently
MySQL and PostgreSQL can be used as database backends.

DBMail is made up of several components. A normal MTA (Postfix,
SendMail, QMail, Exim) is used for accepting messages. The MTA
hands the messages over to dbmail-smtp, using a pipe interface, or
dbmail-lmtpd, using LMTP (Local Mail Transport Protocol). These
programs take care of delivering the message into the database.
Messages can be retreived from the database using dbmail-pop3d,
using the POP3 protocol, and dbmail-imapd, using the IMAP4Rev1
protocol.

The whole email is stored in the database. That includes attachments.
The DBMail programs do not have to touch the filesystem to retreive
or insert emails. User information is also stored in the database,
so users do not need an account on the machines DBMail is running
on.