Penguin Power, version 0.1.5
Ppower, short for 'Penguin Power', is a piece of UNIX software for controlling
x10 home automation equipment connected to the computer via a cm11a interface.
It can be used to turn things on and off and lights dim and bright.  In
addition to controlling the x10 devices, it can receive commands sent to the
x10 devices from other sources such as an x10 remote transceiver and execute
commands on your system in response.

The virtue of ppower is that it is designed to run as a daemon at all times,
keeping track of the status of the X10 system, taking action on events,
queueing requests, and providing information on the events it is receiving to
any interested clients.