Initial import of ruby-ncurses-1.0 from pkgsrc-wip, packaged by Rasputnik.
This ruby extension makes most functions, constants, and external
variables of the C library ncurses accessible from the Ruby
programming language.

All C functions are wrapped by module functions of a the module
"Ncurses", with exactly the same name. Additionally, C functions
expecting a WINDOW* as their first argument can also be called
as methods of the "Ncurses::WINDOW" class.

The panel library (for support of overlapping windows) is also
wrapped, in the module "Ncurses::Panel".