Import dictem version 0.81 from pkgsrc-wip, packaged by Aleksey Cheusov. This is newer version than PR 33203.
DictEm is a customizable DICT client for (X)Emacs. It implements all
functions of the client part of the DICT protocol (RFC-2229). Unlike
dictionary.el, it widely uses autocompletion that is used for
selecting a dictionary and search strategy. It also provides several
hooks that may be used for buffer postprocessing. Built-in
hyperlinking and a highlighting mechanisms are based on this
ability. Also, DictEm supports the mechanism of virtual dictionaries
that can be used for grouping dictionaries from _different_ DICT
servers into the client-side virtual dictionary.

Fore information about dictionary protocol and DICT project, see
http://www.dict.org.