Initial import of whois3 version 3.2.2 into The NetBSD Packages Collection.
This is whois3, a very simple and generic whois client.

This client, unlike the "classic" whois client,
does not check for supported flags at the client side,
except for -h (whois host) and -p (whois port).
The syntax checks for flags are made at the server side.
The default host and port of whois3 is whois.ripe.net
at port 43 (or to be more exact, the tcp port specified
for whois in /etc/services).

Note: This is in fact a "reimport" of an updated version
of package net/ripe-whois which is going to be removed.
Most notable change from the last version of net/ripe-whois
package is the addition of the -4 and -6 switches to select
either IPv4 or IPv6 as transport.