Update to 1.6.1:
1.6.1

New versions of OpenEXR and CTL are now available. This release
fixes a buffer overrun in OpenEXR and a Windows build problem in
CTL, and it removes a few unnecessary files from the .tar.gz
packages.

1.6.0

* Reduced generational loss in B44- and B44A-compressed images.
* Added B44A compression. This is a variation of B44, but with
a better compression ratio for images with large uniform areas,
such as in an alpha channel.
* Bug fixes.

1.5.0

*  OpenEXR supports a new image compression method, called B44.
It has a fixed compression rate of 2.28:1, or 4.57:1 if used
in combination with luminance/chroma encoding. B44-compressed
images can be uncompressed fast enough to support real-time
playback of image sequences.
* The new playexr program plays back moving image sequences.
Playexr is multi-threaded and utilizes the threading capabilities
of the IlmImf library that were introduced in OpenEXR 1.3.0.
The program plays back B44-compressed images with fairly
high-resolution in real time on commodity hardware.
* The playexr program and a new version of the existing exrdisplay
image viewer both support color rendering via color transforms
written in the new Color Transformation Language or CTL. CTL
is not part of OpenEXR; it will be released separately. CTL
support in playexr and exrdisplay is optional; the programs
can be built and will run without CTL.
* In preparation for the release of CTL, OpenEXR has been split
into three separate packages:
  o IlmBase 0.9.0 includes the Half, Iex, Imath and IlmThread
  libraries
  o OpenEXR 1.5.0 includes the IlmImf library, programming
  examples and utility programs such as exrheader or
  exrenvmap
  o OpenEXRViewers 0.9.0 includes the playexr and exrdisplay
  programs
* The "Technical Introduction to OpenEXR" document now includes
a recommendation for storing CIE XYZ pixel data in OpenEXR
files.
* A new "OpenEXR Image Viewing Software" document describes
the playexr and exrdisplay programs. It briefly explains
real-time playback and color rendering, and includes recommendations
for testing if other image viewing software displays OpenEXR
images correctly.
* The OpenEXR sample image set now includes B44-compressed
files and files with CIE XYZ pixel data.